Move mdocContext unit tests into mocha test suite.
Converts the standalone test/unit/mdocContext.test.js (added in PR #219)
into a standard mocha suite at test/mocha/91-mdoc-context.js so it runs
under the common testing system, npm test, and CI. Removes the ad-hoc
test:unit script and the test/unit directory.

Addresses #222.
